backend/pyproject.toml
[project]
name = "voltaros-backend"
version = "0.1.0"
description = "Chaos Engineering Orchestrator using ADK"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click~=8.1.7",
    "cookiecutter~=2.5.0",
    "google-cloud-aiplatform>=1.87.0",
    "rich~=13.7.0",
    "pyyaml~=6.0.1",
    "backoff>=2.2.1",
    "chaoslib~=1.11.0",  # Chaos Toolkit
    "google-cloud-monitoring~=3.11.4",
    "google-cloud-bigquery~=3.25.0",
    "google-cloud-storage~=2.18.2",
    "fastapi~=0.115.0",
    "uvicorn~=0.30.6",
]

[dependency-groups]
dev = [
    "black~=23.11.0",
    "flake8~=6.1.0",
    "isort~=5.12.0",
    "pytest~=7.4.3",
    "pytest-cov~=4.1.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "src", "agents"]
testpaths = ["tests"]
addopts = "-s -v"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(levelname)s - %(message)s"